Rename lightning_wrapper to pl_wrapper
